requirements.txt
streamlit
psycopg2-binary
matplotlib
numpy
seaborn
tabulate
streamlit-lottie
